+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Pause arez when changing the replicant connection, otherwise if the server initiates a close and the client has pending requests then the code will react during subscription purge and generate an error while the connection is null but the state has yet to be updated.
+* Ensure non-cached required type graphs are transmitted to the client.
 
 ### [v6.124](https://github.com/replicant4j/replicant/tree/v6.124) (2022-09-19) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.122...v6.124)